+++ xarray/core/utils.py
@@ -32,6 +32,8 @@ from typing import (
 import numpy as np
 import pandas as pd
 
+from . import dtypes
+
 K = TypeVar("K")
 V = TypeVar("V")
 T = TypeVar("T")
@@ -75,6 +77,23 @@ def maybe_cast_to_coords_dtype(label, coords_dtype):
     if coords_dtype.kind == "f" and not isinstance(label, slice):
         label = np.asarray(label, dtype=coords_dtype)
     return label
+
+
+def maybe_coerce_to_str_type(index, original_coords):
+    """maybe coerce a pandas Index back to a nunpy array of type str
+
+    pd.Index uses object-dtype to store str - try to avoid this for coords
+    """
+
+    try:
+        result_type = dtypes.result_type(*original_coords)
+    except TypeError:
+        pass
+    else:
+        if result_type.kind in "SU":
+            index = np.asarray(index, dtype=result_type)
+
+    return index
 
 
 def safe_cast_to_index(array: Any) -> pd.Index: