+++ doc/whats-new.rst
@@ -42,6 +42,9 @@ Bug fixes
   By `Anderson Banihirwe <https://github.com/andersy005>`_
 - Fix a crash in orthogonal indexing on geographic coordinates with ``engine='cfgrib'`` (:issue:`4733` :pull:`4737`).
   By `Alessandro Amici <https://github.com/alexamici>`_
+- Coordinates with dtype ``str`` or ``bytes`` now retain their dtype on ``reindex``, ``align``,
+  ``concat``, and ``assign``, previously they were cast to an object dtype 
+  (:issue:`2658` and :issue:`4543`) by `Mathias Hauser <https://github.com/mathause>`_. 
 
 Documentation
 ~~~~~~~~~~~~~